using *_kernel.h, *_kernel.cu, *_kernel.cpp hiearchy
